Fixed UI for Savings Preference

+++ lib/models/savings_preferences_model.dart
@@ -31,8 +31,8 @@ class SavingsPreferencesData {
   int? savingPreferenceId;
   String? userEmail;
   String? transactionType;
-  int? percentageValue;
-  int? frequencyValue;
+  double? percentageValue;
+  double? frequencyValue;
   String? savingsPreferenceStartDate;
   String? savingsPreferenceEndDate;
   String? preferenceUpdateDate;
